requirements.txt
langgraph
langchain_openai
langchain_community
python-dotenv
duckduckgo-search